requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.32.0
numpy==2.1.3
pydantic==2.10.2
typing-extensions==4.12.2
jax==0.4.35
jaxlib==0.4.35
scipy==1.14.1
redis==5.0.8
requests==2.32.3
GEOparse==2.0.3
celery==5.4.0
scanpy==1.10.4
anndata==0.10.8
pandas==2.2.3
scikit-learn==1.5.2
faiss-cpu==1.12.0
matplotlib==3.9.2
networkx==3.4.1
pytest==8.3.3
pytest-asyncio==0.24.0
httpx==0.27.2
python-multipart==0.0.9